Fixes issue that oocured if the user did not have hideOnScreenAlertInCombat enabled
Previously, it would not display the alert if the user needed to swap weapons from the result of the in-combat when crossing the final equip/unequip boundary in a partial prismatic dungeon check. Instead, it waited for the combat leave event to be fired and be handled by that associated event.